requirements.txt
flask
Flask-SQLAlchemy
Flask-WTF
flask-login
python-dotenv
typer # for cli